requirements.txt
pymongo
fastapi
certifi
python-dotenv
uvicorn
face_recognition
cmake
dlib
PyQt5
torch
torchvision
torchaudio
pyaudio
numpy
soundfile
transformers
pyttsx3
speechrecognition
geopy
geocoder
google-generativeai
pydantic
faiss-cpu
opencv-python
pydub